[refactor] shared generators.common
